+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/impl/FlowControlAPIGroupClient.java
@@ -18,6 +18,7 @@ package io.fabric8.kubernetes.client.impl;
 import io.fabric8.kubernetes.client.dsl.FlowControlAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.V1beta1FlowControlAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.V1beta2FlowControlAPIGroupDSL;
+import io.fabric8.kubernetes.client.dsl.V1beta3FlowControlAPIGroupDSL;
 import io.fabric8.kubernetes.client.extension.ClientAdapter;
 
 public class FlowControlAPIGroupClient extends ClientAdapter<FlowControlAPIGroupClient> implements FlowControlAPIGroupDSL {
@@ -30,6 +31,11 @@ public class FlowControlAPIGroupClient extends ClientAdapter<FlowControlAPIGroup
   @Override
   public V1beta2FlowControlAPIGroupDSL v1beta2() {
     return adapt(V1beta2FlowControlAPIGroupClient.class);
+  }
+
+  @Override
+  public V1beta3FlowControlAPIGroupDSL v1beta3() {
+    return adapt(V1beta3FlowControlAPIGroupClient.class);
   }
 
   @Override